Queue using multithreading method
Create a class called queue which has an array of integers, and front and rear position values to represent the front and rear of the queue, and two Boolean flags full and empty. Implement two synchronized methods get and put to retrieve next item to consume from the queue and to store the newly produced item into the queue. Set the flags appropriately according to the state of the queue. Create one object of the queue and two threads namely producer thread and consumer thread and share the queue for both the threads. Producer thread should produce items and place it in the queue and the consumer will consume the items one by one from the queue. If the queue is full the producer has to wait until one of the items is consumed by the consumer. Similarly if the queue is empty the consumer has to wait until the producer places at least one item into the queue.